Closer to done
- Separated the FBX loading into to methods, one loading the file and
the other loading the mesh inside of it
- A little bit of the texture can be seen on the model, but it's not
done yet

+++ 3D Engine/GeometryBase.h
@@ -27,7 +27,7 @@ struct ModelConfig
 
 	uint id_indices = 0;
 	uint num_vertices = 0;
-	float* vertices = nullptr;
+	uint* vertices = nullptr;
 
 	uint id_normals = 0;
 	uint num_normals = 0;
